apigw-private-endpoint/python/requirements.txt
radish-bdd==0.13.1
gitpython==3.1.20
netaddr==0.8.0
colorful==0.5.4
filetype==1.0.8
junit-xml==1.9
emoji==1.6.1
lxml==4.6.4
semver==2.13.0
IPython==7.16.1
orjson==3.6.0
diskcache==5.1.0
terraform-compliance==1.3.30
flask==2.0.2
python-dotenv==0.19.2
flask-restful==0.3.9